backend/requirements.txt
Django>=4.2,<5
dj-database-url>=2.1
django-cors-headers>=4.3
djangorestframework>=3.15
gunicorn>=22
psycopg2-binary>=2.9
python-dotenv>=1.0
